+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next release
 
+- Added a test for multiple reservations on the same space slot
 - Fix a security issue: updated moment to 2.29.4 to fix [CVE-2022-31129](https://cve.mitre.org/cgi-bin/cvename.cgi?CVE-2022-31129)
 
 ## v5.4.13 2022 July 27